feat: add TemporalColumnFilter to common
Fixes #462
