+++ psycop/projects/cvd/model_training/populate_cvd_registry.py
@@ -8,7 +8,7 @@ def populate_with_cvd_registry() -> None:
     from psycop.projects.cvd.model_training.preprocessing.regex_filter import (
         RegexColumnBlacklist,
     )
-    from psycop.projects.cvd.model_training.preprocessing.datetime_filter import (
+    from psycop.common.model_training_v2.trainer.preprocessing.steps.col_filters import (
         TemporalColumnFilter,
     )
     from psycop.projects.cvd.model_training.preprocessing.bool_to_int import BoolToInt
